los-agujetas/los_agujetas/requirements.txt
Django==4.2.1       
django-ninja==0.21.0
django-cors-headers==4.0.0